Split up AnimatedImplementation.js
