+++ Libraries/Animated/src/NativeAnimatedHelper.js
@@ -8,6 +8,7 @@
  *
  * @providesModule NativeAnimatedHelper
  * @flow
+ * @format
  */
 'use strict';
 
@@ -15,6 +16,9 @@ const NativeAnimatedModule = require('NativeModules').NativeAnimatedModule;
 const NativeEventEmitter = require('NativeEventEmitter');
 
 const invariant = require('fbjs/lib/invariant');
+
+import type {AnimationConfig} from './animations/Animation';
+import type {EventConfig} from './AnimatedEvent';
 
 let __nativeAnimatedNodeTagCount = 1; /* used for animated nodes */
 let __nativeAnimationIdCount = 1; /* used for started animations */
@@ -49,13 +53,26 @@ const API = {
     assertNativeAnimatedModule();
     NativeAnimatedModule.connectAnimatedNodes(parentTag, childTag);
   },
-  disconnectAnimatedNodes: function(parentTag: ?number, childTag: ?number): void {
+  disconnectAnimatedNodes: function(
+    parentTag: ?number,
+    childTag: ?number,
+  ): void {
     assertNativeAnimatedModule();
     NativeAnimatedModule.disconnectAnimatedNodes(parentTag, childTag);
   },
-  startAnimatingNode: function(animationId: ?number, nodeTag: ?number, config: Object, endCallback: EndCallback): void {
+  startAnimatingNode: function(
+    animationId: ?number,
+    nodeTag: ?number,
+    config: Object,
+    endCallback: EndCallback,
+  ): void {
     assertNativeAnimatedModule();
-    NativeAnimatedModule.startAnimatingNode(animationId, nodeTag, config, endCallback);
+    NativeAnimatedModule.startAnimatingNode(
+      animationId,
+      nodeTag,
+      config,
+      endCallback,
+    );
   },
   stopAnimation: function(animationId: ?number) {
     assertNativeAnimatedModule();
@@ -77,11 +94,17 @@ const API = {
     assertNativeAnimatedModule();
     NativeAnimatedModule.extractAnimatedNodeOffset(nodeTag);
   },
-  connectAnimatedNodeToView: function(nodeTag: ?number, viewTag: ?number): void {
+  connectAnimatedNodeToView: function(
+    nodeTag: ?number,
+    viewTag: ?number,
+  ): void {
     assertNativeAnimatedModule();
     NativeAnimatedModule.connectAnimatedNodeToView(nodeTag, viewTag);
   },
-  disconnectAnimatedNodeFromView: function(nodeTag: ?number, viewTag: ?number): void {
+  disconnectAnimatedNodeFromView: function(
+    nodeTag: ?number,
+    viewTag: ?number,
+  ): void {
     assertNativeAnimatedModule();
     NativeAnimatedModule.disconnectAnimatedNodeFromView(nodeTag, viewTag);
   },
@@ -89,14 +112,30 @@ const API = {
     assertNativeAnimatedModule();
     NativeAnimatedModule.dropAnimatedNode(tag);
   },
-  addAnimatedEventToView: function(viewTag: ?number, eventName: string, eventMapping: EventMapping) {
+  addAnimatedEventToView: function(
+    viewTag: ?number,
+    eventName: string,
+    eventMapping: EventMapping,
+  ) {
     assertNativeAnimatedModule();
-    NativeAnimatedModule.addAnimatedEventToView(viewTag, eventName, eventMapping);
+    NativeAnimatedModule.addAnimatedEventToView(
+      viewTag,
+      eventName,
+      eventMapping,
+    );
   },
-  removeAnimatedEventFromView(viewTag: ?number, eventName: string, animatedNodeTag: ?number) {
+  removeAnimatedEventFromView(
+    viewTag: ?number,
+    eventName: string,
+    animatedNodeTag: ?number,
+  ) {
     assertNativeAnimatedModule();
-    NativeAnimatedModule.removeAnimatedEventFromView(viewTag, eventName, animatedNodeTag);
-  }
+    NativeAnimatedModule.removeAnimatedEventFromView(
+      viewTag,
+      eventName,
+      animatedNodeTag,
+    );
+  },
 };
 
 /**
@@ -128,9 +167,11 @@ const TRANSFORM_WHITELIST = {
 };
 
 function validateTransform(configs: Array<Object>): void {
-  configs.forEach((config) => {
+  configs.forEach(config => {
     if (!TRANSFORM_WHITELIST.hasOwnProperty(config.property)) {
-      throw new Error(`Property '${config.property}' is not supported by native animated module`);
+      throw new Error(
+        `Property '${config.property}' is not supported by native animated module`,
+      );
     }
   });
 }
@@ -138,7 +179,9 @@ function validateTransform(configs: Array<Object>): void {
 function validateStyles(styles: Object): void {
   for (var key in styles) {
     if (!STYLES_WHITELIST.hasOwnProperty(key)) {
-      throw new Error(`Style property '${key}' is not supported by native animated module`);
+      throw new Error(
+        `Style property '${key}' is not supported by native animated module`,
+      );
     }
   }
 }
@@ -153,7 +196,9 @@ function validateInterpolation(config: Object): void {
   };
   for (var key in config) {
     if (!SUPPORTED_INTERPOLATION_PARAMS.hasOwnProperty(key)) {
-      throw new Error(`Interpolation property '${key}' is not supported by native animated module`);
+      throw new Error(
+        `Interpolation property '${key}' is not supported by native animated module`,
+      );
     }
   }
 }
@@ -170,8 +215,24 @@ function assertNativeAnimatedModule(): void {
   invariant(NativeAnimatedModule, 'Native animated module is not available');
 }
 
-function isNativeAnimatedAvailable(): boolean {
-  return !!NativeAnimatedModule;
+let _warnedMissingNativeAnimated = false;
+
+function shouldUseNativeDriver(config: AnimationConfig | EventConfig): boolean {
+  if (config.useNativeDriver && !NativeAnimatedModule) {
+    if (!_warnedMissingNativeAnimated) {
+      console.warn(
+        'Animated: `useNativeDriver` is not supported because the native ' +
+          'animated module is missing. Falling back to JS-based animation. To ' +
+          'resolve this, add `RCTAnimation` module to this app, or remove ' +
+          '`useNativeDriver`. ' +
+          'More info: https://github.com/facebook/react-native/issues/11094#issuecomment-263240420',
+      );
+      _warnedMissingNativeAnimated = true;
+    }
+    return false;
+  }
+
+  return config.useNativeDriver || false;
 }
 
 module.exports = {
@@ -182,7 +243,7 @@ module.exports = {
   generateNewNodeTag,
   generateNewAnimationId,
   assertNativeAnimatedModule,
-  isNativeAnimatedAvailable,
+  shouldUseNativeDriver,
   get nativeEventEmitter() {
     if (!nativeEventEmitter) {
       nativeEventEmitter = new NativeEventEmitter(NativeAnimatedModule);
